+++ base_sms_client/__manifest__.py
@@ -1,18 +1,20 @@
 # coding: utf-8
-# © 2004-2009 OpenERP SA (<http://openerp.com>)
+# Copyright (C) 2015 Sébastien BEAU <sebastien.beau@akretion.com>
 # © 2011 SYLEAM (<http://syleam.fr/>)
 # © 2013 Julius Network Solutions SARL <contact@julius.fr>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 
 {
-    "name": "SMS Client Core",
+    "name": "Base Sms Client",
     "version": "10.0.1.0.0",
-    "depends": ["base",
-                "mail",
+    'license': 'AGPL-3',
+    "depends": ['mail',
                 'base_phone',
                 'keychain',
                 ],
+    'author': 'Julius Network Solutions,SYLEAM,'
+              'Odoo Community Association (OCA),Akretion',
     'images': [
         'images/sms.jpeg',
         'images/gateway.jpeg',
@@ -20,20 +22,16 @@
         'images/client.jpeg',
         'images/send_sms.jpeg'
     ],
-    "summary": """
-Sending SMSs very easily, individually or collectively.
-
-    """,
+    "summary": "Sending SMSs very easily, individually or collectively.",
     "website": "http://julius.fr",
     "category": "Phone",
-    "demo": [],
     "data": [
         "security/groups.xml",
         "security/ir.model.access.csv",
         "security/ir.rule.csv",
         "views/sms_gateway_view.xml",
-        "views/serveraction_view.xml",
-        "views/sms_gateway_data.xml",
+        "views/server_action_view.xml",
+        "data/sms_gateway_data.xml",
         "wizard/mass_sms_view.xml",
         "views/smstemplate_view.xml"
     ],
