+++ base_sms_client/i18n/ar.po
