+++ base_sms_client/i18n/de.po
