+++ base_sms_client/i18n/fr.po
