+++ base_sms_client/security/groups.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+    <record id="group_sms_user" model="res.groups">
+        <field name="name">SMS / User</field>
+    </record>
+</odoo>
