+++ base_sms_client/static/src/img/icon.png
