+++ base_sms_client/views/sms_gateway_view.xml
@@ -1,0 +1,140 @@
+<?xml version="1.0"?>
+<odoo>
+    <record model="ir.ui.view" id="sms_gateway_form">
+        <field name="model">sms.gateway</field>
+        <field name="priority" eval="8"/>
+        <field name="arch" type="xml">
+            <form string="SMS Gateway">
+                <sheet>
+                    <group col="2">
+                        <h1><field name="name" nolabel="1" placeholder="Name" colspan="2"/></h1>
+                        <field name="method" nolabel="1" placeholder="Method" colspan="2"/>
+                        <field name="default_gateway"/>
+                        <field name="company_id" groups="base.group_multi_company"/>
+                    </group>
+                    <notebook colspan="4">
+                        <page string="General">
+                            <group col="4">
+                                <field name="state"/>
+                                <field name="from_provider"/>
+                                <field name="url"/>
+                            </group>
+                        </page>
+                        <page string="Permission">
+                            <separator string="Access Permission"
+                                colspan="4" />
+                            <field name="user_ids" colspan="4" nolabel="1"/>
+                        </page>
+                        <page string="Additional option">
+                            <group>
+                                <group>
+                                    <field name="validity" />
+                                    <field name="classes"/>
+                                    <field name="deferred" />
+                                    <field name="nostop" />
+                                </group>
+                                <group>
+                                    <field name="priority" />
+                                    <field name="coding"/>
+                                    <field name="tag"/>
+                                    <field name="char_limit"/>
+                                </group>
+                            </group>
+                        </page>
+                    </notebook>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="sms_gateway_tree">
+        <field name="model">sms.gateway</field>
+        <field name="priority" eval="8"/>
+        <field name="arch" type="xml">
+            <tree string="SMS Gateways">
+                <field name="name"/>
+                <field name="method"/>
+            </tree>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_sms_gateway_tree">
+        <field name="name">SMS Gateway</field>
+        <field name="res_model">sms.gateway</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form,tree</field>
+        <field name="view_id" ref="sms_gateway_tree" />
+    </record>
+
+    <menuitem name="SMS Gateway"
+              id="menu_gateway_administration_sms_server"
+              sequence="100"
+              parent="base_phone.menu_config_phone"
+              action="action_sms_gateway_tree"/>
+
+    <record model="ir.ui.view" id="sms_sms_tree">
+        <field name="model">sms.sms</field>
+        <field name="arch" type="xml">
+            <tree string="Sms">
+                <field name="create_date"/>
+                <field name="mobile"/>
+                <field name="message"/>
+                <field name="state"/>
+                <field name="gateway_id"/>
+                <button string="Send" states="draft" type="object" name="send"/>
+                <button string="Retry" states="error" type="object" name="retry"/>
+                <button string="Cancel" states="draft,error" type="object" name="cancel"/>
+            </tree>
+        </field>
+    </record>
+
+    <record model="ir.ui.view" id="sms_sms_form">
+        <field name="model">sms.sms</field>
+        <field name="arch" type="xml">
+            <form string="Sms">
+                <header>
+                    <button string="Send" states="draft" type="object" name="send"/>
+                    <button string="Retry" states="error" type="object" name="retry"/>
+                    <button string="Cancel" states="draft,error" type="object" name="cancel"/>
+                    <field name="state" widget="statusbar" statusbar_visible="draft,send"/>
+                </header>
+                <sheet>
+                    <label for="partner_id" class="oe_edit_only"/>
+                    <h1><field name="partner_id"/></h1>
+                    <label for="mobile" class="oe_edit_only"/>
+                    <h1><field name="mobile"/></h1>
+                    <group col="4">
+                        <field name="create_date" readonly="True"/>
+                        <field name="gateway_id" colspan="2"/>
+                        <field name="company_id" groups="base.group_multi_company"/>
+                        <separator string="SMS Message" colspan="4"/>
+                        <field name="message" colspan="4" nolabel="1"/>
+                        <separator string="Last Error" colspan="4"/>
+                        <field name="error" colspan="4" nolabel="1"/>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_sms_sms_tree">
+        <field name="name">Sms</field>
+        <field name="res_model">sms.sms</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form,tree</field>
+        <field name="view_id" ref="sms_sms_tree" />
+    </record>
+
+    <menuitem name="SMS Message"
+              id="menu_gateway_administration_sms"
+              sequence="110"
+              parent="base_phone.menu_config_phone"
+              action="action_sms_sms_tree"/>
+
+    <act_window context="{'gateway_id': active_id}"
+                domain="[('gateway_id', '=', active_id)]"
+                id="act_sms_gateway_2_sms"
+                name="Sms Message"
+                res_model="sms.sms"
+                src_model="sms.gateway"/>
+</odoo>
