+++ base_sms_client/wizard/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2013 Julius Network Solutions SARL <contact@julius.fr>
+#    Author: Sébastien BEAU <sebastien.beau@akretion.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from . import mass_sms
