+++ ovh_sms_client/models/sms_gateway.py
@@ -3,10 +3,10 @@
 # Valentin CHEMIERE <valentin.chemiere@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import api, models
+from odoo import api, models, fields
+from ..models.keychain import OVH_KEYCHAIN_NAMESPACE
 import requests
 import logging
-import json
 
 _logger = logging.getLogger(__name__)
 
@@ -14,22 +14,18 @@ _logger = logging.getLogger(__name__)
 class SmsClient(models.Model):
     _inherit = "sms.gateway"
 
-    @api.model
-    def get_method(self):
-        method = super(SmsClient, self).get_method()
-        method.append(('http_ovh', 'OVH HTTP'), )
-        return method
+    method = fields.Selection(selection_add=[('http_ovh', 'OVH HTTP')])
 
     @api.multi
     def _provider_get_provider_conf(self):
         for rec in self:
             keychain = rec.env['keychain.account']
-            if rec._check_permissions:
+            if rec._check_permissions():
                 retrieve = keychain.suspend_security().retrieve
             else:
                 retrieve = keychain.retrieve
             accounts = retrieve(
-                [['namespace', '=', 'ovh_provider']])
+                [['namespace', '=', OVH_KEYCHAIN_NAMESPACE]])
             return accounts[0]
 
 
@@ -40,9 +36,9 @@ class SmsSms(models.Model):
     def _prepare_http_ovh(self):
 
         keychain_account = self.gateway_id._provider_get_provider_conf()
-        keychain_json = json.loads(keychain_account['data'])
+        keychain_data = keychain_account.get_data()
         params = {
-            'smsAccount': keychain_json['sms_account'],
+            'smsAccount': keychain_data['sms_account'],
             'login': keychain_account['login'],
             'password': keychain_account.get_password(),
             'from': self.gateway_id.from_provider,
@@ -78,7 +74,6 @@ class SmsSms(models.Model):
             'smsAccount': '*****',
             'login': '*****',
             })
-        print params
         _logger.debug("Call OVH API : %s params %s",
                       params['url'], params)
         response = r.text
