--- smsclient_core/views/serveraction_view.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-        <record model="ir.ui.view" id="view_server_statistics_graph">
-            <field name="name">ir.actions.server.form.inherit</field>
-            <field name="model">ir.actions.server</field>
-            <field name="inherit_id" ref="base.view_server_action_form"/>
-            <field name="arch" type="xml">
-                <page name="code" position="after" autofocus="autofocus">
-                    <page string="SMS" name="sms" attrs="{'invisible': [('state', '!=', 'sms')]}">
-                        <group>
-                            <field name="sms"/>
-                            <field name="mobile"/>
-                            <field name="sms_server" invisible="1"/>
-                            <field name="sms_template_id" domain="[('sms_template','=',True)]" attrs="{'required':[('state','=','sms')]}"/>
-                        </group>
-                    </page>
-                </page>
-            </field>
-        </record>
-</odoo>
