--- smsclient_core/views/smstemplate_view.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data>
-
-        <!-- ************************************************************** -->
-        <!-- ** SMS Template ****************************************** -->
-        <!-- ************************************************************** -->
-        <record model="ir.ui.view" id="email_template_form">
-            <field name="name">email.template.form</field>
-            <field name="model">mail.template</field>
-            <field name="inherit_id" ref="mail.email_template_form"/>
-            <field name="arch" type="xml">
-            	<field name="email_to" position="after">
-            		<field name="mobile_to"/>
-				</field>
-				<field name="lang" position="before">
-					<field name="sms_template"/>
-					<field name="gateway_id"/>
-				</field>
-            </field>
-        </record>
-    </data>
-</openerp>
