--- smsclient_core/wizard/mass_sms_view.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="view_wizard_send_sms">
-            <field name="model">wizard.send.sms</field>
-            <field name="arch" type="xml">
-            	<form string="SMS Gateway - Sens SMS" version="7.0">
-	                <group string="Gateway" colspan="4">
-                        <field name="gateway_id"/>
-					</group>
-					<notebook>
-	                    <page string="Message">
-                            <field name="message" colspan="4" nolabel="1"/>
-                            <!-- TODO limit to 10 the number of partner to show -->
-                            <field name="partner_ids" colspan="4" nolabel="1">
-                                <tree options="{'limit': 10}">
-                                    <field name="name"/>
-                                    <field name="mobile"/>
-                                </tree>
-                            </field>
-                        </page>
-	                   	<page string="Additional option">
-   		                	<group>	  	 	
-	           		          	<group>
-	   		          	        	<field name="validity"/>
-	   		          	        	<field name="classes"/>
-	   		          	        	<field name="deferred"/>
-	   		          	        	<field name="nostop"/>
-	           		          	</group>
-	   		          	        <group>
-		                   			<field name="priority"/>
-		                   			<field name="coding"/>
-		                   			<field name="tag"/>
-								</group>
-							</group>
-						</page>
-					</notebook>
-		            <footer>
-		                <button string="Send SMS" name="send" type="object" class="oe_highlight"/>
-		                or
-		                <button string="Cancel" class="oe_link" special="cancel"/>
-		            </footer>
-	            </form>
-            </field>
-        </record>
-        
- 		<act_window name="Send an SMS"
-            res_model="wizard.send.sms"
-            src_model="res.partner"
-            view_mode="form"
-            target="new"
-            key2="client_action_multi"
-            id="action_wizard_send_sms" />
-    </data>
-</openerp>
